Give the Majin-Bo a custom base type
As with the other staff unrands, adjust it to be more appropriate for its
slot, making it more useful as a melee weapon by giving it a custom base
type. It keeps its one-handedness and ends up with 11 base damage, +1 base
accuracy, and 1.2 base delay (the same damage as a demon whip but with 0.1
higher base and min delay).