Add some new Slime Pit endings
slime_pit: The main change was splitting out the various functions for
making the clear walls disappear when TRJ dies. I also removed some
downstairs glyphs, which were in there for some reason. I removed the
"no_dump" tag, since now that there are variants, players might want to
see which one they got. I also trimmed some excess walls off the edges
to get the width less than 70 tiles, allowing the vault to be rotated,
and accordingly removed the "no_rotate" tag.

slime_pit_nicolae_amoeboid: A variation on the basic slime_pit theme --
stairs near the outer edge of slime walls, with stone, loot, and TRJ
in the center.

slime_pit_nicolae_inversion: Switching it around a little: stairs near
an outer edge of stone walls, with loot and other decorative bits around
the outer edge. The center holds slime walls and TRJ. As the comment
notes, there is usually an acidless path to the center, but it's not
guaranteed.

slime_pit_nicolae_jellycorners: Instead of starting near the edge and
working one's way inwards, you start at the center and work your way
outwards. Here, TRJ is lurking in one of the square stone ruins in the
corners. The decaying stone walls might end up providing a little niche
to hide in, but also maybe not.

slime_pit_nicolae_royal_road_to_ruin: For a bit, I considered a layout
that was mostly just an A-to-B path, rather than mostly open areas the
player can wander through; this concept was rejected and the map I'd
made was turned into nicolae_slime_spelunking. The loot/TRJ section from
that was kept and reused here. Here, the three stairs start in a
mostly-open area, and then converge on the center together, turning into
a short A-to-B path. Like slime_pit_nicolae_jellycorners, players will
walk through decaying ruins of the pre-Slime civilization, which may or
may not provide fightin' niches.

slime_pit_nicolae_trefoil: A concept of 3-towards-1, in which there are
three distinct paths to the center, and going from one upstair to another
requires going through the TRJ zone. Here, there's quite a bit more loot,
so there are more monsters to compensate.

For all the new vaults, I mostly kept the same amount and ratios of
monsters (types, numbers) and loot (* relative to |) as in slime_pit.
That stuff is a pain to balance, and I figured it'd be easiest to go
with what I know. I also didn't bother weighting any of the vaults;
that can come later with feedback.