Experimental: give monks an orb of light
There are two thoughts behind this change:
1. A background that starts with an orb of light might feel
   feel interestingly different, insofar as high-EV and invisible
   enemies become substantially less threatening with light held.
2. Monks are a bit weak (as backgrounds go), and light feels
   sufficiently 'divine' to fit with their theme.

I'm a bit worried that light has too much staying power for a starting
item - starting items should generally be replaced or obsolete at some
point, but I can easily see some characters keeping an orb of light
all game. That reduces variety when playing these backgrounds.

But this is cheap and easy to revert, so let's try it!