Adjust the sphere of Battle
Let it enhance all destructive spells instead of only conjurations, but
make it act as a negative enhancer for non-destructive spells (giving it
more of an orb-like downside).

The description implied that the summoned battlesphere's power was based
purely off Conjurations skill, but instead its power was calculated exactly
as if casting the spell (factoring in intelligence and other enhancers).
Change it to only look at Conjurations skill (resulting in it being easier
to reach higher power since no stepdown is applied).