Tweak *Corrode and *Slow artefacts
Instead of each source of *Corrode giving an independent 3% chance to apply
a level of corrosion on taking damage, make sources additive (so the very
unlikely case of stacking lots of sources of *Corrode can no longer apply
multiple levels of corrosion). Do the same for *Slow (which has a 1% chance
per source).