Fix trailing comma for UI JSON (tonyspumoni)
json_write_comma() is intended to be called immediately *before*
writing another object. If there's already a comma or opening brace
or similar as the previous character, then json_write_comma is a no-op.
This nicely (if somewhat confusingly) solves the issue of writing
commas before all but the last element of a list.

Closes #3078.