fix: remove a redundant/broken CI step
Our current invocation of actions/checkout fetches the entire repository
history and the tags, and then this fetch line exists to apparently
shorten the history to depth 1. This isn't normally a big deal, but if
the most recent tag on a branch is non-annotated, this breaks git
describe on that branch. Current trunk is like this because of the
debian tag.

It's possible there's something I'm still not appreciating about what
this stop was intended to accomplish? Both changes to fetch depth were
added in a single commit in 449611fb78e74826.