Potion of flight -> enlightenment
The potion of flight has long been somewhat bland - it's often
worth an inventory slot, but it's never exciting to use. Meanwhile,
we've long had requests for a potion which granted Will, like
resistance's damage effects. What if we solved both problems at
once, and added a new pun at the same time?