Fix the description of walls with unseen changes (take two)
Temporary terrain changes produced by the summon forest spell can
affect dungeon features outside of the player's LOS. Those features
lose the flavour when described or if the dungeon view is reloaded.
The description is also wrong for the summoned trees after the spell
finished.

The first version of the patch had to be reverted because it crashed
on some vaults with loot on top of custom floor.