Reduce BVC base damage
Borgnjor's Vile Clutch was recently buffed by 57a6a0d306af, which let
it work on unseen and out-of-LOS enemies. It was already considered a
strong spell before that change, so focus it more on the distinctive part
of the spell (immobilizing enemies and reducing their EV) by reducing
the damage component. At 50 power (25% of max), the spell goes from
2d6 damage to 2d4.5; at 150 power, it goes from 2d11 to 2d9.5.