Fix "I'll leave it outside" shop message.
When purchasing items from a shop when you don't have sufficient inventory space, a message is supposed to appear that notifies you that the shop keeper will put the items on the ground for you.

This message was not appearing because it was being immediately reset with line 1328 (update_help) before ever getting drawn. I've written a fix that changes the behavior of update_help depending on if the last command caused items to be left outside.