Fix "I'll leave it outside" shop message
When purchasing items from a shop when you don't have sufficient inventory
space, a message is supposed to appear that notifies you that the shop keeper
will put the items on the ground for you.

This message was not appearing because it was being immediately reset with line
1328 (update_help) before ever getting drawn. This fix changes update_help based
on whether the last command caused items to be left outside.

Closes #3149.

[Committer's note: I share the author's concerns about whether this is the
cleanest way to fix the bug, but don't have time to dig further into menu code.]