Ensmallen the inventory at the Tiny Tomes shop
It was pointed out on Discord that with 20 books, it can take an
annoyingly long time to look through all of them to examine their
contents. Accordingly, dial it down a bit by making it symmetric
with Tremendous Tomes, with each shop having 8 books.
(I have other ideas for shops with the maximum number of items,
anyway. (watch this space)))