feat: separator escaping for certain rc options
This implements escaping separators for certain rc options, most
importantly force_more_message and message_colour. There is so much
demand for this that many standard rc files (including crawl itself, see
e.g. 7dd7712201) have assumed this is already implemented, but until
0.30, there were no error messages to indicate when this was going
wrong.

For many (but not all) list options, this commit will let `,` be escaped
via `\,`. This really only matters for options where arbitrary regexes
are used. For the two options mentioned above, this commit also
implements escaping `:` via `\:`. The behavior for force_more_message
where an invalid channel prefix causes a `:` split to be ignored is
retained. (These are implemented slightly more generally, via the
message_filter and colour_mapping objects, but I don't think the
escaping matters for any other cases.)

This is a bit of a hodgepodge because of various different
implementations of separators, so I have not tried to make the change
very general.  Because the `runrest...` options are implemented in lua,
it is unfortunately not practical to implement escaping even `,` without
a major rework. Also, it would be easy to add this for options that use
split_parse, but as far as I can tell none really need it.