feat: show scale when zooming in local tiles
It's good to show explicit feedback to the user for this (previously
only debug builds got this). This might be better to show somewhere else
in the UI, but for now it's easiest to just do in MSGCH_PROMPT, and
seems harmless.

(cherry picked from commit 2ddbf375f0fc33f2fc56da2359d7c675bdcd68b0)