build: workaround for an actions/checkout bug
a9a6e9e7af1 and ea43df2a32 removed a step that was probably intended to
work around https://github.com/actions/checkout/issues/290 and
https://github.com/actions/checkout/issues/882, but broke the CI build
if anyone used a lightweight tag (because these fetch lines actually
shallow the tag depth to 1). This commit attempts to thread the needle
by making this workaround conditional only to releases, where (if the
person doing the release has followed the steps correctly), there is
guaranteed to be an annotated tag at depth 1 in the tag history. (If I'm
parsing the checkout bugs correctly, they only trigger on a tag action
anyways.)