refactor: treat first option name as canonical
This is a variant on a change due to Aliscans in https://github.com/crawl/crawl/pull/2873/commits/0db8a29bacb10f4a431f44341df5cdce4f245b65.

This commit takes the proposed API change there, but leaves the
implementation as a set for lookup purposes. Currently the canonical
name isn't used, but it will be in future commits.