Turn off quiver autoswitching (CarefulOdds, advil)
Switching to a new quivered item when you ran out of the thing you
were shooting made sense back when folks were juggling +2 elvish arrows
and +1 arrows of flame. These days, though, it's pretty uncommon that
the player wants to keep chucking stuff when their quiver runs empty,
and much more likely to be disastrous if e.g. the quiver switches from
boomerangs to scrolls of fear.

Let's disable this and see if anyone complains.