feat: generalized string conversion for ListGameOption
General conversion to string for GameOption subclasses has long been a
desirable feature.  #2873 also needs string conversion to work right,
and it contains two not-quite-general attempts for ListGameOption. The
first one used `<<` overloads, and seems not to have done what was
needed (I had a hard time decoding why exactly this was dropped).
Additionally, the << code was a bit awkward, especially in the context
of crawlcode where it isn't particularly idiomatic. (Personally, I
despise the stream API, so I won't need much convincing not to go this
route.) The second attempt subclasses ListGameOption for two special
cases, and has cleaner code for those two cases, but drops the attempt
at generality.

This commit takes a different approach. C++11 added the std::to_string
function for basic types, and while this isn't itself supposed to be
overloaded (caveat: debated and unclear, also we do it in a few places)
this change, in combination with ADL and namespaces, provides a
reasonably clear template for how general string conversion can be done
a mostly clean way. This commit implements the options::to_string
function at a variety of overloads and template specializations, and
uses it to implement general string serialization for ListGameOption.
(In principle, this might be used to generalize a whole lot of other
code, but this commit doesn't do that for now.)

1. If a std::string overload exists, it will be used. (Because of the
history of things, this is currently just unused and provided for
generality. But this may later simplify some other things...)
2. If a type defines an appropriate function ::str(), it will be used.
3. A few special case enums are handled (e.g. `job_type`). These are
very easy to write.
4. An overload for `std::string` is provided (mysteriously, not part of
`std::to_string`)
5. If none of the above handle the case of interest, code will error on
a deleted `to_string` specialization.

2 uses some wacky SFINAE stuff but I think it's moderately cleaner than
my last attempt at this..