refactor: simplify some ListGameOption SFINAE code
This removes the more baroque ListGameOption parts of 12f39dd2a6 in
favor of an `options::from_string` templated call for specific types in
ListGameOptions, as a counterpart to the recently added
`options::to_string`. This has the disdvantage of not being able to use
arbitrary conversion functions any more (you can only write one at any
given type), but has massively cleaner code and no macros needed. For
now, I didn't try to relocate the constructor-based SFINAE into
options::from_string.

As part of this I revamped an undocumented sound option to be a little
more sane. I'm moderately unclear on why this option exists or what
exactly it does, the doc additions are based on how it appears in
initfile.cc.