Feat: magic mapping -> revelation
Magic mapping is a cool item, but I often find it sitting in my
inventory unused for the great majority of a game. I might use it
early on to find a timed portal if I'm having trouble, but after
that, it sits around until Zot. That's not exciting!

Meanwhile, the removal of Ashenzari's Astral Vision in 1891117ce2a
(Sept 2022) left a mechanical gap. We removed the "see through walls"
effect for a variety of UI and bug reasons, but it's still neat.
If the player only saw through walls for a moment, until the end of
their turn, there'd be no problem. So... let's kill two birds with one
scroll!

Scrolls of revelation still map the level, like magic mapping, but
they also reveal everything in LOS radius - monsters, items, unseen
horrors, you name it. This should hopefully create more incentive
to use magic mapping throughout the game - you might want to see
what's inside a scary looking vault or behind a door, for example.

It does break the legendary door vault a bit, but truthfully, that
thing was meant to be broken. If players want to spend a scarce
resource to peek inside, they can do that. :)

UI issue - invisible monsters vanish completely once your turn ends,
so you might want to note them down. Might poke at this at some point.