refactor: remove old monster glyph and keybind files
Added after a significant overhaul of monster glyphs and colours in 0.4
(f2f6755ac), monster_glyphs files allowed to make the glyphs of all
affected monsters show up as they did in previous versions.

There doesn't seem to be much demand for such files now, so it's time to
remove them.

Also, this commit removes the file with 0.3.4's key bindings that
somehow survived in the codebase for 15 years.