feat: Don't let the Damnation unrand harm the shooter
As of a99462a43, Damnation requires hitting an actual monster in order
to explode. Hence the player has no way to safely use the weapon on an
adjacent monster without harming themself. This creates an annoying need
to swap to a backup launcher frequently, since even when you have
exploding damnation bolts, monsters get adjacent quite often.

Hence this commit makes Damnation simply not affect the shooter. It's a
buff to Damnation, but the nerf in 404d6944 was strong, and the weapon
isn't overpowered these days. I also removed some ammo mulching code
that was specific to Damnation but is no longer necessary when launcher
ammo always mulches.