Bugfix: Fixed autofight_nomove_fires option being ignored
Due to the erroneous condition, autofight_nomove_fires option was ignored
if autofight_fires was set to true. Added a function
should_quiver_be_used(no_move), which returns true if the quiver may
be used based on the current settings and action. Let me know if this name
is unsuitable.