docs: document a secret X-mode command
Added in af5862b9e, `CMD_MAP_EXCLUDE_RADIUS` allows placing exclusions
with arbitrary sizes on the map. The command was supposed to be
available only in the wiz-mode, but due to a missing `you.wizard` check
anyone can use it.

Since the command might be useful for checking the range of monsters'
spells or inner-flame explosions, let's document it in a few relevant
places.