fix: convert a sanctuary ASSERT to a check
This is somewhat theoretical, but, I think if a sanctuary ends exactly
when changing level, and the coords that were the center of the
sanctuary are in los on the new level, there's a sequencing issue in
loading the new level where env.sanctuary_pos is reset, but the FPROPs
aren't yet cleared (I'm 95% sure they do get cleared later in the level
load process). This case triggered a crash on `is_sanctuary`.

This commit tweaks things so that this simply isn't a crash, we don't
check the FPROPs unless sanctuary_pos is set. This in principle would
prevent identifying a class of bugs with the FPROPs not getting cleared,
but I haven't run across such a bug in many years.

Resolves #3170

(cherry picked from commit bc41564c69dc6bdc9b55f8742df79b2f87626955)