feat: A wizard command to unobtain runes and Orb
There's no existing convenient way to unget runes or the Orb of Zot, and
for testing purposes it's nice to be able to do this. The ideal UI would
present a toggle for each rune and the Orb, but that's more work to
implement and it's easy enough to re-obtain runes/orb with the wizard
mode item commands.