fix: Use the correct type for the minor version
When reading the "chr" section, pass the read minor version as
`uint32_t` and not `uint8_t`, since we converted minor version to the
former in 43576957. Using the wrong type was truncating the minor
version to 0, resulting in some erroneous unmarshalling that corrupted
the value of `you.explore`, among possibly other things.

Note: This fix was cherry-picked from ca8d4b91 in the talisman-why-not
experimental branch, where the bug was noticed, but thankfully never
triggered in trunk.

Also remove an unused tag function and clean up a lack of whitespace.