Add a new Slime Pit ending
In my earlier set of Slime Pit endings, I had one ending that was
intended to be easier than normal, with a lot of corridors. It turned
out to be even easier than I thought thanks to the abundance of
glass walls: TRJ could see the player but wouldn't move around the
walls to get to them, instead staying where it could see them.
When I merged the vaults, I dropped that one ending, thinking I would
get around to making another corridor-heavy vault some other time.
I wasn't expecting "some other time" to be "a day or two later",
but inspiration strikes when it wants. Anyway:
slime_pit_nicolae_block_by_block: Features a number of corridors with
the slime walls mostly confined to the edges, and more narrow areas
as you get closer to the center.