Add a base AC penalty to Blade Hands
Before talismans, blade hands offered a tradeoff: incredible killing
power in exchange for terrible defenses. If you wanted to get blade
hands online ASAP, you probably had to wear very light armour. That
was exciting, and so was the feeling of 'cheating' that restriction
as your skill, Str and Int increased and as you became able to wear
heavier armour along with your very sharp hands. (The above is my
hazy recollection, from the days of Blade Hands being in the Tm
starting book.)

Talismans remove that tradeoff, since they don't care about
encumbrance. For most forms, that's irrelevant (because they merge
armour) or fine and desirable (for beast/death form). But Blade Hands
wants something more.

So: remove the UC base damage scaling from Blade Hands. Replace it
with a penalty to base AC, similar to that given by the 'deformed
body' mutation', which reduces the AC offered by body armour. At
min Shapeshifting skill, this reduces base AC to 0, and at 20 skill,
the penalty is removed entirely. This has no effect on armour
enchantment or on the effects of Armour skill, which still scale
along with the base armour.

This should hopefully provide a similar dynamic to the one I outline
above. Let's try it!