Add talismans to more uniques
For thematic appropriateness and to add more loot to uniques which
are otherwise a bit lacking. Rupert gets a maw talisman (for shoutin'),
Aizul gets a serpent talisman (because she's a snake), and Sojobo gets
a storm talisman (for zzaping.)

Some of these should perhaps be chances instead of guarantees, to avoid
increasing talisman count too much. (In particular, I'm suspicious of
Rupert and Sojobo.)

It'd also be fun to add a blade and/or dragon talisman somewhere.