feat: Tweaks to a Lair ending (elliptic)
For hellmonk_lair_end_abyssal_woods, don't try to place Mlioglotl with a
rare chance. For branch ends, we generally don't place `8` level threats
since the end vault is both important and already has a high difficulty.
We try to keep difficulty a bit more consistent for these maps, and
Mlioglotl is a fast and very hard hitting monster. Due to a slight bug
in my DES when I modified this vault that wasn't caught when the unique
was added to it in ae8ea5ec392b, Mlioglotl couldn't actually generate,
but let's just make this official.

Also replace the ancient zymes in this vault with a raiju. The zyme
sickness mechanic really only works in the Abyss, where having no regen
for any amount of time is impactful and resting or moving is a
non-trivial decision. Although the damage profile of a zyme is fine,
raiju is quite similar, is also an Abyss native, and its unique mechanic
is fine for Lair.