Play with unique talismans
Add dragon-blood talisman to Bai Suzhen (she was on my short-list
initially, and was a popular suggestion from others, e.g. elliptic),
and reduce the odds of some other talismans to a coinflip. Aizul
stays guaranteed (since that serpent talisman is more flavor than
value at the point it appears - until we upgrade it to a randart?!),
and Roxanne is of course iconic.