Fully map the Temple on entering
Adds a new flag for fully mapped branches, and refactors magic mapping to
handle full mapping with a parameter instead a special-cased radius check.

Fully mapped floors provide no exploration healing or piety, and have no
exploration traps (which was already true for the Temple except for the
healing).

It would be nice to also map Gauntlets on entry but there are some extra
things to resolve there - as implemented, monsters and items aren't mapped,
but displaying them would also reveal the minotaur's treasure immediately
(and removing Meteoran exploration healing would be a bit more meaningful
in Gauntlet than in Temple).