feat: Tweak a Shoals vault
For nicolae_shoals_abyss_on_the_beach, have loot placed near the Abyss
portal instead of on top of it and the Lugonu altar, since this way it's
easier to see the loot items at a glance. Add one more star_item to the
loot, bringing the total to three items, so that this runed door vault
is a bit more worthwhile to enter. This very slightly decreases the
average monster placement but still gives it a 2:1 ratio for monsters to
loot.