query item for property
Adds a function which checks the value of the property associated with an artp provided by an item.

This is distinct from artefact_property() in that it is unified across all item types; for example Fire Dragon Scales of Fire Resistance would not be a valid input to artefact_property(), where this method would return the number of pips of rF provided.

The primary use case I have in mind for this is for querying STR or DEX provided by an item, however it could also simplify monster resistances and likely uselessness checks.

This method is less efficient than a purpose built query for e.g. fire resistance, due to the lookup tables for finding the equivalent SPARM and armor flag.