feat:Talisman vault (#3211)
Small vault that places a talisman and corresponding monsters/scenery.

[Committer's note: Limited the sets per branch in terms of threat level and
branch theme, so there's no autoexploring into a Lair dancing weapon or
a Mines air elemental + raiju band. Cleaned up the header, added more of a
border, a door, and the patrolling tag so they wander away a lot less from
the vault theming, since most of these are pretty out-of-depth. Made threat
monsters not sit on top of the talisman so one doesn't miss the reason to
bother with the vault. Give each of them a small chance of being a randart.

Closes #3211.]