Tweak the center of nicolae_shoals_lighthouse
Now that the phantom mirror is unique, there's a good chance it will
be replaced with a wand, which would make the flavor not work.
(Lighthouses use mirrors to reflect the light where it needs to go,
you see.) The only thing I can think of to replace it with is an orb
of light, but honestly, it might just be best to not have any loot
in there in the first place and close it off. So I'll do that.

This also fixes the issue mentioned in this reddit thread:
https://www.reddit.com/r/dcss/comments/152vcev/