Remove vamp from unusual default (acrobat, Ge0ff)
Trying to err on the side of less-is-more with this initial set of commits:
Each additional brand that we add by default here adds noise and dilutes the
impact of the unusual highlighting when it does show up.

For anyone who would really prefer more warnings, please remember you can
still add them to your rc file!