Don't place fire vortices as monsters
As with ball lightnings, they have the "conjured" flag and aren't treated
as proper monsters in many ways, so shouldn't be placed by vaults or as
lava spawns. They remain only in a few cosmetic-only vaults and in some
Sprints, and have been replaced either with fire elementals or with nothing
in other cases.