Remove the Awaken Earth spell
It was intended as a counter to players digging tunnels in Elf, and while
it sort of works in that sense it also acts unintuitively and completely
unlike any other spell, being castable by monsters out of LOS and breaking
into and out of vaults. Give deep elf elementalists LRD instead.