Add a new custom book shop
Back in 2012, I discovered a bug*. You couldn't define custom spellbooks
with more than one spell in a shop. This is because the symbol for
separating items in a shop definition was the same as the symbol for
separating spells in a randbook's spell list: |. The map parser would
just get confused.

In 2015, I submitted a patch that changed the symbol for separating
spells in a spell list to &, allowing custom shops to handle custom
spellbooks. I said, "I promise to abuse this new vault-writing power as
soon as possible." Apparently, "as soon as possible" was "eight years
later".

Some of the book themes are obvious, some less so. No doubt players will
enjoy trying to figure them out. I may have gone slightly overboard, but
that just means more variety for players. I do it all for them.

* https://crawl.develz.org/mantis/view.php?id=5706