Simplify unusual monster markings (Implojin, advil)
To reduce noise and simplify option handling, let's only mark as
unusual the items that are most likely to be relevant to players.
Wands, distortion and chaos weapons, darts and throwing nets are
all marked 'unusual' by default. Artefacts, brands to which the
player is 'vulnerable', spectral weapons, javelins, and items with
plusses are not.

As before, this can be configured further in the player's rcfile.