Appropriately adjust agape appellations (PleasingFungus)
In 37a08a6, PF fixed a "spaces in a shop spec" bug by simply using
a pre-set name rather than a randomly-generated name. This commit
puts the randomly-generated names back in and makes sure they don't
break the spec with spaces.

For this, I simply got rid of the spaces rather than replacing them
with underscores: having a second word in the name that probably
doesn't begin with "A" breaks the alliteration. (Ditching the spaces
is a much easier solution than looking for spaces and changing all
the letters right after them to "A".)