A Shoals entry and an Orc decorative vault (#3230)
 * Four pools, three merfolk or a tougher monster, one per pool.
 * A kraken in a very tiny pool in an oval room.

[Committer's notes: Heavy header clean-up. The first of these has been
altered for far less plants to let autoexplore actually find the Shoals
entry, and swaps out a blind autoexploring into a sea snake in Lair
for a manticore instead. The plants might hypothetically provide some
warning, but if other monsters roar right beside the walls the vault
monsters are likely to tear right through the plants.

The second of these was a kraken Shoals entrance, blocked off by plants.
There's something to be said for how kraken don't really feel much
intimidating any more these days, but there's too much that's awkward
with the use of such in Lair: it's not actually super plausible for a
Lair character to deal enough damage at full range to get the free XP, but
it's both tempting and very tedious to try, plus one can actually just get
mangled doing it. For want of another obvious use for it, the vault's name
vaguely points it at putting it in some mines instead, as one of a million
glass vaults; they didn't kill it and neither should you. Could use
another vault PR to go guarantee Orc:1 decor, maybe?...

Closes #3230.]