fix: don't clear repeats when a monster destroys a floor stack
`dec_mitm_item_quantity` assumed it was being called by player actions
only, but it can be used in a few other cases, notably when a monster
(such as a jiyva slime) destroys a floor stack. I couldn't find a
reliable general way to check whether a player action is responsible for
the call, and this function isn't called that much, so I handled every
case. These clear calls are still necessary for floor actions; without
them a repeat that uses up a scroll stack (for example) with another
stack in the same position has very weird behavior.

This fixes a few other call types that shouldn't mess with command
history, but they probably don't matter as much.

Resolves #3194