Switch the CSET_ASCII character for talismans back to '|'.
In CSET_ASCII, '%' is a very bad choice because it is used for corpses.
Using a different character for menus and for the item on the map isn't
ideal either, but seems less bad.

Ideally there would be an option to fully hide corpses, freeing up '%' for
an item class.