Add tile for player under Glaive of Prune curse (#3239)
It would be helpful for some visual to see when the prune curse kicks in,
& adds fun flavor.

[Committer's notes: Decided on one of the submissions that didn't have any
icy effects on it, due to either coming off as closer to fire or being very
obviously cut off, and desaturated it a bit from the Glaive of Prune's
colours to try and get closer to the specific dark glossiness of a prune.
Implemented actually displaying this with the tech introduced for a31ef7c,
and cleared out a redundant tile while doing so. It's a little weird
looking (though prunes are not the most photogenic of food in the first
place), but the visuals when it happens are 100% worth it, so I am grateful
for the submission.

Closes #3239.]