feat: prompt popup consistency for tiles
Several menus that can appear in the message pane support using popups,
but for historical reasons would do it only when necessary (e.g. there
is already a popup on the stack). The two main cases are the Y/N(/A)
menu, and the travel menu. Crawl taking input in the message pane has a
long roguelike-y tradition, but it prevents mouse input, and can't work
in all cases anyways. This commit moves the default towards consistently
using a popup in UIs where the popup works well. In particular, in local
tiles, and webtiles, these two popups (where the Y/N(/A) is used quite
widely) will now always pop up by default. As part of this change, I
have attempted to make a bunch of the behavior more consistent. Specific
notes:

* For those who prefer the old ways, set `prompt_menu = false`. This is
  the default still for console (where popups still take over the whole
  screen), including on webtiles builds. We'll see how much people hate
  this as a webtiles default...but I think it is a clear improvement on
  local tiles.
* Handling of case in defaults is changed; I think the behavior this is
  addressing is essentially buggy. Previously, for the popup (but not
  the message pane?) if a default was lowercased, but the menu required
  uppercase, the default wouldn't work.
* enter (and locally, space) now work to indicate a default in the popup
* Y/N prompts now have a smaller min width in webtiles
* With menu_arrow_control=false, the defaults are hinted more clearly in
  the popup
* Error messaging in the Y/N popup happens in the more, not in a dummy
  entry
* Arrow selection is now allowed for uppercase Y/N menus