refactor: move altar travel into TravelPromptMenu
This commit refactors some extremely janky prompt menu stuff for travel
to altars into the more structured TravelPromptMenu as a new mode. Like
the regular travel menu, it can appear in either the message pane or a
popup, and it has general and option-dependent code for deciding on
which it tries and how it lays out. I had considered just removing this
submenu entirely (since ctrl-f seems like a better approach to me), but
it sounds like players do use this primarily via the ctrl-o _ path.