Rescue an arrival vault from oblivion and add it
orjb1 submitted a PR with an arrival vault in it:
https://github.com/crawl/crawl/pull/3246
but after initial feedback they seem to have deleted it.
I liked the vault, though, and yoinked it from oblivion.
I made a few changes to get it working right, but other
than that, I kept it mostly as it was submitted. (I didn't
fix the asymmetry of the empty floor spaces around the
edges, even though I mentioned it on the PR; I ended up
liking how it interacted with the rest of the level.)