Remove Gozag's passive gold detection
It was a cute minor flavour effect, but had weird consequences in practice -
granting Gozag a pseudo-mapping ability that was particularly spoilery
around vaults that placed lots of gold. Gold is still moved to the top of
item piles, to keep some of the flavour and make gold still stand out.