Revert "Let monsters know if the player has clarity (Lici)"
Having monsters waste their time with useless spells is intended in some
cases - gaining certain resistances shouldn't make monsters more dangerous
by changing their AI and causing them to only cast spells that the player
is vulnerable to. There's very little consistency with this across
different monster spells, but in various cases the intended behaviour is
for them not to cast spells that the player is "obviously" (by virtue of
species etc) immune to, but to still cast them if the player has gained
resistance/immunity in some other way.

This reverts commit 993fcd7dae2e2a45acc3f264f09929cf83240ccc.