feat: Auto-travel to traversable squares next to non-traversable ones
Added UI improvement that makes it so if the player selects a
non-traversable target with a traversable cell adjacent, prompt before
failing and allow the user to go to the closest adjacent cell instead.

This feature doesn't affect controls that wouldn't have invoked the
"Sorry, I don't know how to traverse that place." msg.
This is intentional. While using touch controls or within the current
view, it should be following the current idea to not spam the user with
prompts and messages on easy misclicks.

This feature could also be expanded to have an options for always
saying yes to the prompt, increased radius for the check (and
corresponding option for it).

This feature request originated from the wiki on UI Improvements