feat: Tweak a Shoals vault (elliptic)
For nicolae_shoals_mertower, rework the design a bit so it's less likely
to be an autoexplore trap into a big chamber of top end merfolk. Have
the monsters be patrolling (appropriate for a tower) and put windows in
the tower as well. Add three pieces of loot, one of which has a 50%
chance to be a tower shield (this is a pun). For the inner monsters,
thin them out so we only place about 3.5 high end merfolk on average,
with javelineers made less common. The average number of monsters placed
is still about 7.

Also remove the tower's inner water. The original intent was for the
tower to be raised above the tides, but not so much that it was dry.
However this is a bit awkward thematically with the outside dry land,
and the vault is plenty hard without guaranteed water.