Fix a crash with confused drowned souls
Confused monsters can get an attack on the player and then another attack
on themselves in the same action. If a drowned soul did this, it would
self-destruct on the first attack, but then try to make the second attack
against themselves before having properly been reset and counted as dead.
Fix by cleaning up dead attacking monsters at the end of a melee attack.