Fix: don't let death-form grow roots (acrobat)
Undead players aren't supposed to be able to transform. They *are*
allowed to leave death form and enter talisman forms, for convenience,
but they shouldn't be able to become trees. Even if Gozag offers!

I considered banning lig from potpet when in death form, but that
seemed inconsistent with other effects that are presently unhelpful.
Sometimes you get a potion you didn't want!