Feat: hot pursuit (elliptic)
Not to be confused with the similar-sounding apparel beloved of
certain players, 'hot pursuit' is a new mechanic to encourage fun
lines of play.

A Brief History
---------------

When in a very nasty spot (low on HP and next to a tough foe), players
have historically been able to 'pillar-dance', wasting time (both the
game's and theirs) to get time to heal. This is both unfun to do and
narratively unsatisfying. When in a tight spot, players should pull out a
cool trick (a spell, god ability or consumable), fight, and/or die!

We first tried to fix this by adding 'random energy', which unfortunately
fixed nothing. Then we tried 'attacks of opportunity', letting monsters
attack when the player moved away. These worked somewhat, but had several
disadvantages, including:

- They were very complex. The list of special cases for which monsters
  could attack the player and when was very long, and it was hard for
  players to track.
- They were very binary. If a monster was next to the player, danger was
  vastly higher than if they were 2 tiles away. If a monster was as fast
  as the player, danger was vastly higher than if they were just a bit
  slower.
- They had odd and unintuitive interactions with polearms (which didn't
  launch reaching attacks of opportunity).
- They were frustrating. Players felt profoundly unhappy when they were
  killed by attacks of opportunity - it felt like the game becoming more
  hostile.

So, let's try something a bit gentler

Hot Pursuit
-----------

When the player moves away from a monster, if that monster then moves
toward the player, they have a one in ten chance of putting on a 'burst
of speed', moving ~25% faster (move delay 0.8) for the next ~20ish turns.
This speed bonus affects the move that triggered it, so players walking
away from an adjacent yak have a 2% chance of getting a surprise bap
from them.

The intent is to again discourage 'pillar-dancing' and other fiddly
stalling tactics (e.g. running across the entire level to get to stairs)
in a 'softer', fuzzier way, without the hard binaries of attacks of
opportunity.

Wu Jian martial manuevers and Serpent's Lash again give immunity to
this effect.

Let's try it out!