Properly warn the player when aiming explosions at empty tiles near allies
Proper stop attack prompts were only being given if the explosion effect was
aimed directly at a creature and NOT if you were aiming at an empty space near
that creature.

This seems to have been a result of the beam code detecting that we would only
hit our ally AFTER the 'target' we were aiming at, and then aiming at the ground
at that spot and assuming that was good enough to spare them. (hint: it was not)

So now, let's just disable suppressing prompts if the source of the damage is
an explosion.