The return of randomized energy (when pursuing the player).
Randomized energy (essentially monsters having a random choice between
0.9/1.0/1.1 move delay) was removed in c2095565. It had two main problems
- the first was that it felt surprising and 'unfair' when an ogre two
steps away suddenly moved and hit the player on the same turn, and the
second was that it incentivized running away from an adjacent monster
until random energy made a space between the two of you (usually only
taking 0-2 hits in the process).

This commit re-implements randomized energy, but only when the monster
is pursuing the player (replacing the experimental swiftness on pursuit).
This hopefully addresses the first issue with randomized energy, since
an ogre will only be able to move + attack if the player was adjacent to it
and just moved away - essentially an attack of opportunity.

The second issue will be addressed by bringing back attacks of opportunity
in a separate commit, making running away from an adjacent monster far
more deadly.

The hope is that randomized energy will help smooth out some of the sharp
edges on attacks of opportunity - we'll try it out.